+++ web/middleware/database/database.go
@@ -2,19 +2,22 @@ package database
 
 import (
 	"bufio"
+	"database/sql/driver"
 	"fmt"
 	"log"
 	"os"
+	"reflect"
+	"regexp"
+	"strconv"
 	"time"
+	"unicode"
 
-	"bitbucket.org/insamo/mvc/datasource"
-	_ "bitbucket.org/insamo/mvc/dialect/mssql"
-	"bitbucket.org/insamo/mvc/web/bootstrap"
+	"github.com/insamo/mvc/datasource"
+	_ "github.com/insamo/mvc/dialect/mssql2008"
+	"github.com/insamo/mvc/web/bootstrap"
 	"github.com/jinzhu/gorm"
-	//_ "github.com/jinzhu/gorm/dialects/mssql"
-	_ "github.com/jinzhu/gorm/dialects/sqlite"
 	_ "github.com/jinzhu/gorm/dialects/postgres"
-	//_ "github.com/denisenkom/go-mssqldb"
+	_ "github.com/jinzhu/gorm/dialects/sqlite"
 
 	"github.com/kataras/golog"
 )
@@ -71,7 +74,8 @@ func Configure(b *bootstrap.Bootstrapper) {
 
 		// Set logger if configured
 		if b.DatabaseLogFile != nil {
-			db.SetLogger(log.New(b.DatabaseLogFile, "", 0))
+			l := DatabaseLogger{log.New(b.DatabaseLogFile, "\r\n", 0)}
+			db.SetLogger(l)
 		}
 
 		// Loading queries
@@ -86,4 +90,114 @@ func Configure(b *bootstrap.Bootstrapper) {
 
 		b.TxFactory[instance] = datasource.NewTransactionFactory(db, queries)
 	}
+}
+
+var (
+	defaultLogger            = DatabaseLogger{log.New(os.Stdout, "\r\n", 0)}
+	sqlRegexp                = regexp.MustCompile(`\?`)
+	numericPlaceHolderRegexp = regexp.MustCompile(`\$\d+`)
+)
+
+var NowFunc = func() time.Time {
+	return time.Now()
+}
+
+func isPrintable(s string) bool {
+	for _, r := range s {
+		if !unicode.IsPrint(r) {
+			return false
+		}
+	}
+	return true
+}
+
+var DatabaseLogFormatter = func(values ...interface{}) (messages []interface{}) {
+	if len(values) > 1 {
+		var (
+			sql             string
+			formattedValues []string
+			level           = values[0]
+			currentTime     = NowFunc().Format("2006-01-02 15:04:05")
+			source          = fmt.Sprintf("(%v)", values[1])
+		)
+
+		messages = []interface{}{source, currentTime}
+
+		if level == "sql" {
+			// duration
+			messages = append(messages, fmt.Sprintf(" [%.2fms] \n", float64(values[2].(time.Duration).Nanoseconds()/1e4)/100.0))
+			// sql
+
+			for _, value := range values[4].([]interface{}) {
+				indirectValue := reflect.Indirect(reflect.ValueOf(value))
+				if indirectValue.IsValid() {
+					value = indirectValue.Interface()
+					if t, ok := value.(time.Time); ok {
+						formattedValues = append(formattedValues, fmt.Sprintf("'%v'", t.Format("2006-01-02 15:04:05")))
+					} else if b, ok := value.([]byte); ok {
+						if str := string(b); isPrintable(str) {
+							formattedValues = append(formattedValues, fmt.Sprintf("'%v'", str))
+						} else {
+							formattedValues = append(formattedValues, "'<binary>'")
+						}
+					} else if r, ok := value.(driver.Valuer); ok {
+						if value, err := r.Value(); err == nil && value != nil {
+							formattedValues = append(formattedValues, fmt.Sprintf("'%v'", value))
+						} else {
+							formattedValues = append(formattedValues, "NULL")
+						}
+					} else {
+						formattedValues = append(formattedValues, fmt.Sprintf("'%v'", value))
+					}
+				} else {
+					formattedValues = append(formattedValues, "NULL")
+				}
+			}
+
+			// differentiate between $n placeholders or else treat like ?
+			if numericPlaceHolderRegexp.MatchString(values[3].(string)) {
+				sql = values[3].(string)
+				for index, value := range formattedValues {
+					placeholder := fmt.Sprintf(`\$%d([^\d]|$)`, index+1)
+					sql = regexp.MustCompile(placeholder).ReplaceAllString(sql, value+"$1")
+				}
+			} else {
+				formattedValuesLength := len(formattedValues)
+				for index, value := range sqlRegexp.Split(values[3].(string), -1) {
+					sql += value
+					if index < formattedValuesLength {
+						sql += formattedValues[index]
+					}
+				}
+			}
+
+			messages = append(messages, sql)
+			messages = append(messages, fmt.Sprintf(" [%v] ", strconv.FormatInt(values[5].(int64), 10)+" rows affected or returned "))
+		} else {
+			messages = append(messages, "")
+			messages = append(messages, values[2:]...)
+			messages = append(messages, "")
+		}
+	}
+
+	return
+}
+
+type DatabaseLoggerInterface interface {
+	Print(v ...interface{})
+}
+
+// LogWriter log writer interface
+type LogWriter interface {
+	Println(v ...interface{})
+}
+
+// Logger default logger
+type DatabaseLogger struct {
+	LogWriter
+}
+
+// Print format & print log
+func (databaseLogger DatabaseLogger) Print(values ...interface{}) {
+	databaseLogger.Println(DatabaseLogFormatter(values...)...)
 }