+++ web/middleware/jwt/jwt.go
@@ -1,7 +1,7 @@
 package jwt
 
 import (
-	"bitbucket.org/insamo/mvc/web/bootstrap"
+	"github.com/insamo/mvc/web/bootstrap"
 
 	"github.com/dgrijalva/jwt-go"
 	jwtmiddleware "github.com/iris-contrib/middleware/jwt"