+++ web/middleware/recover/recover.go
@@ -1,7 +1,7 @@
 package recover
 
 import (
-	"bitbucket.org/insamo/mvc/web/bootstrap"
+	"github.com/insamo/mvc/web/bootstrap"
 	"github.com/kataras/iris/middleware/recover"
 )